test: extract MockProjectRepository
