+++ internal/app/shared_test.go
@@ -7,7 +7,6 @@ import (
 	"gomander/internal/config/domain"
 	"gomander/internal/event"
 	"gomander/internal/eventbus"
-	projectdomain "gomander/internal/project/domain"
 )
 
 type MockLogger struct {
@@ -48,38 +47,6 @@ func (m *MockUserConfigRepository) GetOrCreate() (*domain.Config, error) {
 
 func (m *MockUserConfigRepository) Update(config *domain.Config) error {
 	args := m.Called(config)
-	return args.Error(0)
-}
-
-type MockProjectRepository struct {
-	mock.Mock
-}
-
-func (m *MockProjectRepository) GetAll() ([]projectdomain.Project, error) {
-	args := m.Called()
-	return args.Get(0).([]projectdomain.Project), args.Error(1)
-}
-
-func (m *MockProjectRepository) Get(id string) (*projectdomain.Project, error) {
-	args := m.Called(id)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*projectdomain.Project), args.Error(1)
-}
-
-func (m *MockProjectRepository) Create(project projectdomain.Project) error {
-	args := m.Called(project)
-	return args.Error(0)
-}
-
-func (m *MockProjectRepository) Update(project projectdomain.Project) error {
-	args := m.Called(project)
-	return args.Error(0)
-}
-
-func (m *MockProjectRepository) Delete(id string) error {
-	args := m.Called(id)
 	return args.Error(0)
 }
