+++ internal/project/application/usecases/editproject_test.go
@@ -9,12 +9,13 @@ import (
 
 	"gomander/internal/project/application/usecases"
 	projectdomain "gomander/internal/project/domain"
+	"gomander/internal/project/domain/test"
 )
 
 func TestDefaultEditProject_Execute(t *testing.T) {
 	t.Run("Should edit a project successfully", func(t *testing.T) {
 		// Arrange
-		mockProjectRepository := new(MockProjectRepository)
+		mockProjectRepository := new(test.MockProjectRepository)
 		sut := usecases.NewEditProject(mockProjectRepository)
 
 		project := projectdomain.Project{Id: "1", Name: "A", WorkingDirectory: "/a"}
@@ -30,7 +31,7 @@ func TestDefaultEditProject_Execute(t *testing.T) {
 
 	t.Run("Should return an error if project update fails", func(t *testing.T) {
 		// Arrange
-		mockProjectRepository := new(MockProjectRepository)
+		mockProjectRepository := new(test.MockProjectRepository)
 
 		sut := usecases.NewEditProject(mockProjectRepository)
